% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa-sequences.R
\name{get_aa_sequences}
\alias{get_aa_sequences}
\title{List AA sequences}
\usage{
get_aa_sequences(aa_sequence_id = NULL)
}
\arguments{
\item{aa_sequence_id}{aa_sequences id}
}
\value{
data from aa_sequences endpoint
}
\description{
AA Sequences are the working units of cells that make everything run (they help make structures, catalyze reactions and allow for signaling - a kind of internal cell communication). On Benchling, these are comprised of a string of amino acids and collections of other attributes, such as annotations.
}
