% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_app-configuration-items.R
\name{get_app_configuration_items}
\alias{get_app_configuration_items}
\title{Get app configuration items}
\usage{
get_app_configuration_items(item_id = NULL, ...)
}
\arguments{
\item{item_id}{app-configuration-items id}

\item{...}{additional query parameters; see Details below}
}
\value{
data from \code{app_configuration_items} endpoint
}
\description{
Create and manage Benchling apps on your tenant
}
\details{
Available query parameters:
\describe{
\item{nextToken}{Token for pagination}
\item{pageSize}{Number of results to return. Defaults to 50, maximum of 100.}
\item{modifiedAt}{Datetime, in RFC 3339 format. Supports the > and < operators. Time zone defaults to UTC. Restricts results to those modified in the specified range. e.g. > 2017-04-30.}
\item{appId}{App id to which the configuration belongs.}
\item{ids}{Comma-separated list of ids. Matches all of the provided IDs, or returns a 400 error that includes a list of which IDs are invalid.}
}
}
