% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{get_apps}
\alias{get_apps}
\title{List apps}
\usage{
get_apps(app_id = NULL)
}
\arguments{
\item{app_id}{apps id}
}
\value{
data from apps endpoint
}
\description{
Create and manage Benchling apps on your tenant
}
