% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-result-schemas.R
\name{get_assay_result_schemas}
\alias{get_assay_result_schemas}
\title{List assay result schemas}
\usage{
get_assay_result_schemas(schema_id = NULL)
}
\arguments{
\item{schema_id}{assay_result_schemas id}
}
\value{
data from assay_result_schemas endpoint
}
\description{
Schemas represent custom configuration of objects in Benchling. See this \href{https://docs.benchling.com/docs/schemas}{guide in our documentation} on how Schemas
impact our developers
}
