% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-results.R
\name{get_assay_results}
\alias{get_assay_results}
\title{List results}
\usage{
get_assay_results(assay_result_id = NULL)
}
\arguments{
\item{assay_result_id}{assay_results id}
}
\value{
data from assay_results endpoint
}
\description{
Results represent the output of assays that have been performed. You can
customize the schemas of results to fit your needs. Results can link to runs,
batches, and other types.To learn more about creating results, \href{https://docs.benchling.com/docs/example-creating-results}{click here}.
}
