% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-runs.R
\name{get_assay_runs}
\alias{get_assay_runs}
\title{List runs}
\usage{
get_assay_runs(assay_run_id = NULL)
}
\arguments{
\item{assay_run_id}{assay_runs id}
}
\value{
data from assay_runs endpoint
}
\description{
Runs capture the details / parameters of a run that was performed. Results are
usually nested under a run.
}
