% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_automation-file-transforms.R
\name{get_automation_file_transforms}
\alias{get_automation_file_transforms}
\title{Lab Automation}
\usage{
get_automation_file_transforms(transform_id = NULL, ...)
}
\arguments{
\item{transform_id}{automation_file_transforms id}

\item{...}{(ignored)}
}
\value{
data from \code{automation_file_transforms} endpoint
}
\description{
Lab Automation endpoints support integration with lab instruments, and liquid
handlers to create samples or results, and capture transfers between containers
at scale.
}
