% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation-input-generators.R
\name{get_automation_input_generators}
\alias{get_automation_input_generators}
\title{Lab Automation}
\usage{
get_automation_input_generators(input_generator_id = NULL)
}
\arguments{
\item{input_generator_id}{automation_input_generators id}
}
\value{
data from automation_input_generators endpoint
}
\description{
Lab Automation endpoints support integration with lab instruments, and liquid
handlers to create samples or results, and capture transfers between containers
at scale.
}
