% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_automation-output-processors.R
\name{get_automation_output_processors}
\alias{get_automation_output_processors}
\title{List non-empty Automation Output Processors}
\usage{
get_automation_output_processors(output_processor_id = NULL, ...)
}
\arguments{
\item{output_processor_id}{automation_output_processors id}

\item{...}{additional query parameters; see Details below}
}
\value{
data from \code{automation_output_processors} endpoint
}
\description{
Lab Automation endpoints support integration with lab instruments, and liquid
handlers to create samples or results, and capture transfers between containers
at scale.
}
\details{
Available query parameters:
\describe{
\item{assayRunId}{Id of the Run}
\item{automationFileConfigName}{Name of the Automation File Config}
\item{archiveReason}{Archive reason. Restricts items to those with the specified archive reason. Use "NOT_ARCHIVED" to filter for unarchived processors. Use "ANY_ARCHIVED" to filter for archived processors regardless of reason. Use "ANY_ARCHIVED_OR_NOT_ARCHIVED" to return items for both archived and unarchived.}
\item{modifiedAt}{Datetime, in RFC 3339 format. Supports the > and < operators. Time zone defaults to UTC. Restricts results to those modified in the specified range. e.g. > 2017-04-30.}
\item{nextToken}{Token for pagination}
}
}
