% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-schemas.R
\name{get_batch_schemas}
\alias{get_batch_schemas}
\title{List batch schemas}
\usage{
get_batch_schemas(schema_id = NULL)
}
\arguments{
\item{schema_id}{batch_schemas id}
}
\value{
data from batch_schemas endpoint
}
\description{
Schemas represent custom configuration of objects in Benchling. See this \href{https://docs.benchling.com/docs/schemas}{guide in our documentation} on how Schemas impact our developers
}
