% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batches.R
\name{get_batches}
\alias{get_batches}
\title{List batches}
\usage{
get_batches(batch_id = NULL)
}
\arguments{
\item{batch_id}{batches id}
}
\value{
data from batches endpoint
}
\description{
Batches represent physical quantities of a sequence, protein, or other biological entity. A batch is identified by its biological entity's name (or Registry ID, if the entity is registered), its date of creation, and an optional inventory number.
}
