% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blobs.R
\name{get_blobs}
\alias{get_blobs}
\title{Blobs}
\usage{
get_blobs(blob_id = NULL)
}
\arguments{
\item{blob_id}{blobs id}
}
\value{
data from blobs endpoint
}
\description{
Blobs are opaque files that can be linked to other items in Benchling, like
assay runs or results. For example, you can upload a blob, then upload an assay
result that links to that blob by ID. The blob will then appear as part of the
assay result in the Benchling web UI.
}
