% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{get_boxes}
\alias{get_boxes}
\title{List boxes}
\usage{
get_boxes(box_id = NULL)

get_boxes(box_id = NULL)
}
\arguments{
\item{box_id}{boxes id}
}
\value{
data from boxes endpoint

data from boxes endpoint
}
\description{
Boxes are a structured storage type, consisting of a grid of positions that can each hold one container. Unlike locations, there are a maximum number of containers that a box can hold (one per position).Boxes are all associated with schemas, which define the type of the box (e.g. "10x10 Cryo Box") along with the fields that are tracked and the dimensions of the box.Like all storage, every Box has a barcode that is unique across the registry.

Boxes are a structured storage type, consisting of a grid of positions that can each hold one container. Unlike locations, there are a maximum number of containers that a box can hold (one per position).Boxes are all associated with schemas, which define the type of the box (e.g. "10x10 Cryo Box") along with the fields that are tracked and the dimensions of the box.Like all storage, every Box has a barcode that is unique across the registry.
}
