% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-entities.R
\name{get_custom_entities}
\alias{get_custom_entities}
\title{List custom entities}
\usage{
get_custom_entities(custom_entity_id = NULL)
}
\arguments{
\item{custom_entity_id}{custom_entities id}
}
\value{
data from custom_entities endpoint
}
\description{
Benchling supports custom entities for biological entities that are neither sequences or proteins. Custom entities must have an entity schema set and can have both schema fields and custom fields.
}
