% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna-alignments.R
\name{get_dna_alignments}
\alias{get_dna_alignments}
\title{List DNA Alignments}
\usage{
get_dna_alignments(dna_alignment_id = NULL)
}
\arguments{
\item{dna_alignment_id}{dna_alignments id}
}
\value{
data from dna_alignments endpoint
}
\description{
A DNA alignment is a Benchling object representing an alignment of multiple DNA sequences.
}
