% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna-oligos.R
\name{get_dna_oligos}
\alias{get_dna_oligos}
\title{List DNA Oligos}
\usage{
get_dna_oligos(oligo_id = NULL)
}
\arguments{
\item{oligo_id}{dna_oligos id}
}
\value{
data from dna_oligos endpoint
}
\description{
DNA Oligos are short linear DNA sequences that can be attached as primers to
full DNA sequences. Just like other entities, they support schemas, tags, and
aliases.
}
