% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna-sequences.R
\name{get_dna_sequences}
\alias{get_dna_sequences}
\title{List DNA sequences}
\usage{
get_dna_sequences(dna_sequence_id = NULL)
}
\arguments{
\item{dna_sequence_id}{dna_sequences id}
}
\value{
data from dna_sequences endpoint
}
\description{
DNA sequences are the bread and butter of the Benchling Molecular Biology
suite. On Benchling, these are comprised of a string of nucleotides and
collections of other attributes, such as annotations and primers.
}
