% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_dropdowns.R
\name{get_dropdowns}
\alias{get_dropdowns}
\title{List dropdowns}
\usage{
get_dropdowns(dropdown_id = NULL, ...)
}
\arguments{
\item{dropdown_id}{dropdowns id}

\item{...}{additional query parameters; see Details below}
}
\value{
data from \code{dropdowns} endpoint
}
\description{
Dropdowns are registry-wide enums. Use dropdowns to standardize on spelling and
naming conventions, especially for important metadata like resistance markers.
}
\details{
Available query parameters:
\describe{

}
}
