% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{get_entities}
\alias{get_entities}
\title{Batches}
\usage{
get_entities(entity_id = NULL)
}
\arguments{
\item{entity_id}{entities id}
}
\value{
data from entities endpoint
}
\description{
Batches represent physical quantities of a sequence, protein, or other
biological entity. A batch is identified by its biological entity's name (or
Registry ID, if the entity is registered), its date of creation, and an
optional inventory number.
}
