% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-schemas.R
\name{get_entity_schemas}
\alias{get_entity_schemas}
\title{List entity schemas}
\usage{
get_entity_schemas(schema_id = NULL)
}
\arguments{
\item{schema_id}{entity_schemas id}
}
\value{
data from entity_schemas endpoint
}
\description{
Schemas represent custom configuration of objects in Benchling. See this \href{https://docs.benchling.com/docs/schemas}{guide in our documentation} on how Schemas impact our developers
}
