% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{get_entries}
\alias{get_entries}
\title{List entries}
\usage{
get_entries(entry_id = NULL)
}
\arguments{
\item{entry_id}{entries id}
}
\value{
data from entries endpoint
}
\description{
Entries are rich text documents that allow you to capture all of your
experimental data in one place.
}
