% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry-schemas.R
\name{get_entry_schemas}
\alias{get_entry_schemas}
\title{List entry schemas}
\usage{
get_entry_schemas(schema_id = NULL)
}
\arguments{
\item{schema_id}{entry_schemas id}
}
\value{
data from entry_schemas endpoint
}
\description{
Schemas represent custom configuration of objects in Benchling. See this \href{https://docs.benchling.com/docs/schemas}{guide in our documentation} on how Schemas
impact our developers
}
