% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entry-templates.R
\name{get_entry_templates}
\alias{get_entry_templates}
\title{List entry templates}
\usage{
get_entry_templates(entry_template_id = NULL)
}
\arguments{
\item{entry_template_id}{entry_templates id}
}
\value{
data from entry_templates endpoint
}
\description{
Entries are rich text documents that allow you to capture all of your experimental data in one place.
}
