% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_events.R
\name{get_events}
\alias{get_events}
\title{List Events}
\usage{
get_events()
}
\value{
data from \code{events} endpoint
}
\description{
The Events system allows external services to subscribe to events that are
triggered in Benchling (e.g. plasmid registration, request submission, etc).
}
