% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{extract_entity_fields}
\alias{extract_entity_fields}
\title{Extract entity data fields}
\usage{
extract_entity_fields(entity_id)
}
\arguments{
\item{entity_id}{entity id or object returned from \code{\link[=get_custom_entities]{get_custom_entities()}}}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} representing the entity data and relevant classes
}
\description{
Extract entity data fields
}
\examples{
\dontrun{
ent_id <- get_custom_entities()$id[1]
extract_entity_fields(ent_id)
get_custom_entities(ent_id) |>
  extract_entity_fields()
}
}
