% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{extract_entity_schema_definition}
\alias{extract_entity_schema_definition}
\title{Extract entity schema definition}
\usage{
extract_entity_schema_definition(entity, quiet = FALSE)
}
\arguments{
\item{entity}{entity id or object returned from \code{\link[=get_custom_entities]{get_custom_entities()}}}

\item{quiet}{should the entity and schema IDs be printed?}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} representing the schema
}
\description{
Extract entity schema definition
}
\examples{
\dontrun{
ent_id <- get_custom_entities()$id[1]
extract_entity_schema_definition(ent_id)
get_custom_entities(ent_id) |>
   extract_entity_schema_definition()
}
}
