% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{extract_schema_definition}
\alias{extract_schema_definition}
\title{Extract schema definition}
\usage{
extract_schema_definition(schema_id)
}
\arguments{
\item{schema_id}{schema id or object returned from \code{\link[=get_entity_schemas]{get_entity_schemas()}}}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} representing the schema
}
\description{
Extract schema definition
}
\examples{
sch_id <- get_entity_schemas()$id[1]
extract_schema_definition(sch_id)
get_entity_schemas(sch_id) |>
   extract_schema_definition()
}
