% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_feature-libraries.R
\name{get_feature_libraries}
\alias{get_feature_libraries}
\title{List Feature Libraries}
\usage{
get_feature_libraries(feature_library_id = NULL, ...)
}
\arguments{
\item{feature_library_id}{feature-libraries id}

\item{...}{additional query parameters; see Details below}
}
\value{
data from \code{feature_libraries} endpoint
}
\description{
Feature Libraries are collections of shared canonical patterns that can be used
to generate annotations on matching regions of DNA Sequences or AA Sequences.
}
\details{
Available query parameters:
\describe{
\item{pageSize}{Number of results to return. Defaults to 50, maximum of 100.}
\item{nextToken}{Token for pagination}
\item{modifiedAt}{Datetime, in RFC 3339 format. Supports the > and < operators. Time zone defaults to UTC. Restricts results to those modified in the specified range. e.g. > 2017-04-30.}
\item{name}{Name of a Feature Library. Restricts results to those with the specified name.}
\item{nameIncludes}{Name substring of a Feature Library. Restricts results to those with names that include the provided substring.}
\item{ids}{Comma-separated list of ids. Matches all of the provided IDs, or returns a 400 error that includes a list of which IDs are invalid.}
\item{names.anyOf}{Comma-separated list of names. Restricts results to those that match any of the specified names.}
}
}
