% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_api_yaml}
\alias{get_api_yaml}
\title{Fetch API Specification}
\usage{
get_api_yaml(org = Sys.getenv("BENCHLING_ORG", ""))
}
\arguments{
\item{org}{benchling tenant. Use empty string (\code{org = ""}) to use the public
API spec \url{https://benchling.com/api/v2/openapi.yaml}}
}
\value{
API spec as a nested list
}
\description{
Fetch API Specification
}
