% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{get_attachments}
\alias{get_attachments}
\title{Get Note Entry Attachments}
\usage{
get_attachments(entry_id, annotateURL = TRUE)
}
\arguments{
\item{entry_id}{entries id}

\item{annotateURL}{should the \code{downloadURL} and \code{size} be annotated? This
requires an additional query}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} containing the name of the attachment (\code{text}),
\code{externalFileId}, and if \code{annotateURL}, the \code{downloadURL} (for the given
user - containing API keys) and \code{size}
}
\description{
Get Note Entry Attachments
}
