% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{get_box_contents}
\alias{get_box_contents}
\title{List box contents}
\usage{
get_box_contents(box_id = NULL, ...)
}
\arguments{
\item{box_id}{box id}

\item{...}{additional query parameters; see Details below

Available query parameters:
\describe{
\item{pageSize}{Number of results to return. Defaults to 50, maximum of 100.}
\item{nextToken}{Token for pagination}
}}
}
\value{
data from \code{boxes/{box_id}/contents} endpoint
}
\description{
Boxes are a structured storage type, consisting of a grid of positions that can
each hold one container. Unlike locations, there are a maximum number of
containers that a box can hold (one per position).Boxes are all associated with
schemas, which define the type of the box (e.g. "10x10 Cryo Box") along with
the fields that are tracked and the dimensions of the box.Like all storage,
every Box has a barcode that is unique across the registry.
}
