% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{get_dropdown_id_from_item_name}
\alias{get_dropdown_id_from_item_name}
\title{Get a dropdown id from an item}
\usage{
get_dropdown_id_from_item_name(dropdown, item)
}
\arguments{
\item{dropdown}{id or name of a dropdown}

\item{item}{name of an item in \code{dropdown}}
}
\value{
id of the \code{item} entry in \code{dropdown}
}
\description{
Get a dropdown id from an item
}
