% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_custom_entity.R
\name{get_dropdown_listing}
\alias{get_dropdown_listing}
\title{Get dropdown listing}
\usage{
get_dropdown_listing(dropdown)
}
\arguments{
\item{dropdown}{name or id of a dropdown}
}
\value{
a two-column \code{data.frame} containing the ids and names of the items
in \code{dropdown}
}
\description{
Get dropdown listing
}
