% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{get_entry_by_id}
\alias{get_entry_by_id}
\title{Get Note Entry Using ID}
\usage{
get_entry_by_id(displayId = NULL)
}
\arguments{
\item{displayId}{displayId of note entry (starting with \code{EXP})}
}
\value{
data from \code{entries} endpoint. For more flexibility, use \code{\link[=get_entries]{get_entries()}}
}
\description{
Get Note Entry Using ID
}
