% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds.R
\name{get_org}
\alias{get_org}
\title{Get the Configured Benchling Organisation}
\usage{
get_org()
}
\value{
the organisation prefix to be used in API calls
}
\description{
Get the Configured Benchling Organisation
}
\details{
if \code{\link[=set_creds]{set_creds()}} has not yet been invoked, it will be with the
default inputs
}
