% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_custom_entity.R
\name{get_schema_data_by_name}
\alias{get_schema_data_by_name}
\title{Get schema data using schema name}
\usage{
get_schema_data_by_name(schema)
}
\arguments{
\item{schema}{name of a schema}
}
\value{
a data.frame of schema definition
}
\description{
Get schema data using schema name
}
