% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{get_schema_id_by_name}
\alias{get_schema_id_by_name}
\title{Get schema id using schema name}
\usage{
get_schema_id_by_name(schema)
}
\arguments{
\item{schema}{name of a schema}
}
\value{
a \code{list} with schema name, registryID, and id
}
\description{
Get schema id using schema name
}
