% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{get_tables}
\alias{get_tables}
\title{Get Note Entry Tables}
\usage{
get_tables(entry_id)
}
\arguments{
\item{entry_id}{entries id}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} containing the \code{name} of the table,
\code{columnLabels}, and \code{rows}, nested appropriately
}
\description{
Get Note Entry Tables
}
\details{
Note that this cannot return the contents of a \code{lookup_table}.

Expand the contents of a table with \code{\link[=extract_table]{extract_table()}}, e.g.

\if{html}{\out{<div class="sourceCode">}}\preformatted{x <- get_tables(entry_id)
extract_table(x[1, ])
}\if{html}{\out{</div>}}
}
