% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds.R
\name{get_token}
\alias{get_token}
\title{Get the Configured Benchling User API Token}
\usage{
get_token()
}
\value{
the user API token to be used in API calls
}
\description{
Get the Configured Benchling User API Token
}
\details{
if \code{\link[=set_creds]{set_creds()}} has not yet been invoked, it will be with the
default inputs
}
