% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_legacy-workflow-stage-runs.R
\name{get_legacy_workflow_stage_runs}
\alias{get_legacy_workflow_stage_runs}
\title{Legacy Workflows}
\usage{
get_legacy_workflow_stage_runs(stage_run_id = NULL, ...)
}
\arguments{
\item{stage_run_id}{legacy_workflow_stage_runs id}

\item{...}{(ignored)}
}
\value{
data from \code{legacy_workflow_stage_runs} endpoint
}
\description{
Legacy workflows allow orchestrating complex experiments.
}
