% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_legacy-workflow-stages.R
\name{get_legacy_workflow_stages}
\alias{get_legacy_workflow_stages}
\title{Legacy Workflows}
\usage{
get_legacy_workflow_stages(stage_id = NULL, ...)
}
\arguments{
\item{stage_id}{legacy_workflow_stages id}

\item{...}{(ignored)}
}
\value{
data from \code{legacy_workflow_stages} endpoint
}
\description{
Legacy workflows allow orchestrating complex experiments.
}
