% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-workflows.R
\name{get_legacy_workflows}
\alias{get_legacy_workflows}
\title{List legacy workflows}
\usage{
get_legacy_workflows(legacy_workflow_id = NULL)
}
\arguments{
\item{legacy_workflow_id}{legacy_workflows id}
}
\value{
data from legacy_workflows endpoint
}
\description{
Legacy workflows allow orchestrating complex experiments.
}
