% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_location-schemas.R
\name{get_location_schemas}
\alias{get_location_schemas}
\title{List location schemas}
\usage{
get_location_schemas(schema_id = NULL, ...)
}
\arguments{
\item{schema_id}{location-schemas id}

\item{...}{additional query parameters; see Details below}
}
\value{
data from \code{location_schemas} endpoint
}
\description{
Schemas represent custom configuration of objects in Benchling. See this \href{https://docs.benchling.com/docs/schemas}{guide in our documentation} on how Schemas
impact our developers
}
\details{
Available query parameters:
\describe{

}
}
