% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{get_locations}
\alias{get_locations}
\title{List locations}
\usage{
get_locations(location_id = NULL)
}
\arguments{
\item{location_id}{locations id}
}
\value{
data from locations endpoint
}
\description{
Manage locations objects.Like all storage, every Location has a barcode that is unique across the registry.
}
