% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtures.R
\name{get_mixtures}
\alias{get_mixtures}
\title{List mixtures}
\usage{
get_mixtures(mixture_id = NULL)
}
\arguments{
\item{mixture_id}{mixtures id}
}
\value{
data from mixtures endpoint
}
\description{
Mixtures are solutions comprised of multiple ingredients where the exact quantities of each ingredient are important to track. Each ingredient is uniquely identified by its component entity.
}
