% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oligos.R
\name{get_oligos}
\alias{get_oligos}
\title{List Oligos}
\usage{
get_oligos(oligo_id = NULL)
}
\arguments{
\item{oligo_id}{oligos id}
}
\value{
data from oligos endpoint
}
\description{
Oligos are short linear DNA sequences that can be attached as primers to full DNA sequences. Just like other entities, they support schemas, tags, and aliases.Please migrate to the corresponding DNA Oligos endpoints so that we can support RNA Oligos.
}
