% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{get_organizations}
\alias{get_organizations}
\title{List organizations}
\usage{
get_organizations(organization_id = NULL)
}
\arguments{
\item{organization_id}{organizations id}
}
\value{
data from organizations endpoint
}
\description{
View organization objects.
}
