% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registries.R
\name{get_registries}
\alias{get_registries}
\title{List registries}
\usage{
get_registries(registry_id = NULL)
}
\arguments{
\item{registry_id}{registries id}
}
\value{
data from registries endpoint
}
\description{
List label templates.
List printers.
Manage registry objects.See our documentation on \href{https://docs.benchling.com/docs/registering-entities}{how to register entities}.
}
