% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-fulfillments.R
\name{get_request_fulfillments}
\alias{get_request_fulfillments}
\title{List Request Fulfillments}
\usage{
get_request_fulfillments(request_fulfillment_id = NULL)
}
\arguments{
\item{request_fulfillment_id}{request_fulfillments id}
}
\value{
data from request_fulfillments endpoint
}
\description{
Requests allow scientists and teams to collaborate around experimental assays and workflows.
}
