% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-schemas.R
\name{get_request_schemas}
\alias{get_request_schemas}
\title{List request schemas}
\usage{
get_request_schemas(schema_id = NULL)
}
\arguments{
\item{schema_id}{request_schemas id}
}
\value{
data from request_schemas endpoint
}
\description{
Schemas represent custom configuration of objects in Benchling. See this \href{https://docs.benchling.com/docs/schemas}{guide in our documentation} on how Schemas
impact our developers
}
