% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{get_requests}
\alias{get_requests}
\title{List requests}
\usage{
get_requests(request_id = NULL)
}
\arguments{
\item{request_id}{requests id}
}
\value{
data from requests endpoint
}
\description{
Requests allow scientists and teams to collaborate around experimental assays
and workflows.
}
