% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna-oligos.R
\name{get_rna_oligos}
\alias{get_rna_oligos}
\title{List RNA Oligos}
\usage{
get_rna_oligos(oligo_id = NULL)
}
\arguments{
\item{oligo_id}{rna_oligos id}
}
\value{
data from rna_oligos endpoint
}
\description{
RNA Oligos are short linear RNA sequences that can be attached as primers to
full DNA sequences. Just like other entities, they support schemas, tags, and
aliases.
}
