% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tasks.R
\name{get_tasks}
\alias{get_tasks}
\title{Tasks}
\usage{
get_tasks(task_id = NULL, ...)
}
\arguments{
\item{task_id}{tasks id}

\item{...}{(ignored)}
}
\value{
data from \code{tasks} endpoint
}
\description{
Endpoints that perform expensive computations launch long-running tasks. These
endpoints return the task ID (a UUID) in the response body.After launching a
task, periodically invoke the \href{#/Tasks/getTask}{Get a task} endpoint with the
task UUID (e.g., every 10 seconds), until the status is no longer RUNNING.You
can access a task for up to 30 minutes after its completion, after which its
data will no longer be available.
}
