% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_custom_entity.R
\name{update_custom_entity}
\alias{update_custom_entity}
\title{Update custom-entity}
\usage{
update_custom_entity(
  custom_entity_id = NULL,
  name = NULL,
  aliases = NULL,
  authorIds = NULL,
  customFields = NULL,
  fields = NULL,
  folderId = NULL,
  entityRegistryId = NULL,
  dry_run = TRUE,
  ...
)
}
\arguments{
\item{custom_entity_id}{registry ID of entity to update}

\item{name}{name of entity to be set}

\item{aliases}{alias of entity to be set}

\item{authorIds}{author ID(s) of entity to be set}

\item{customFields}{custom Fields of entity to be set (as a named list)}

\item{fields}{fields of entity to be set (as a named list)}

\item{folderId}{folder of entity to be set}

\item{entityRegistryId}{registry ID of entity to be set}
}
\value{
if \code{dry_run == TRUE} the response from the server, otherwise \code{FALSE}
}
\description{
Update custom-entity
}
