% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-outputs.R
\name{get_workflow_outputs}
\alias{get_workflow_outputs}
\title{List workflow outputs}
\usage{
get_workflow_outputs(workflow_output_id = NULL)
}
\arguments{
\item{workflow_output_id}{workflow_outputs id}
}
\value{
data from workflow_outputs endpoint
}
\description{
Workflow outputs are outputs of a workflow task
}
