% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-stage-runs.R
\name{get_workflow_stage_runs}
\alias{get_workflow_stage_runs}
\title{Legacy Workflows (deprecated)}
\usage{
get_workflow_stage_runs(stage_run_id = NULL)
}
\arguments{
\item{stage_run_id}{workflow_stage_runs id}
}
\value{
data from workflow_stage_runs endpoint
}
\description{
Please use endpoints for Legacy Workflows. These deprecated endpoints will be
removed once users are migrated onto Legacy Workflows endpoints.
}
