% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-task-groups.R
\name{get_workflow_task_groups}
\alias{get_workflow_task_groups}
\title{List workflow task groups}
\usage{
get_workflow_task_groups(workflow_task_group_id = NULL)
}
\arguments{
\item{workflow_task_group_id}{workflow_task_groups id}
}
\value{
data from workflow_task_groups endpoint
}
\description{
Workflow task groups are groups of workflow tasks of the same schema
}
