% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{get_workflows}
\alias{get_workflows}
\title{List workflows}
\usage{
get_workflows(workflow_id = NULL)
}
\arguments{
\item{workflow_id}{workflows id}
}
\value{
data from workflows endpoint
}
\description{
Please use endpoints for Legacy Workflows. These deprecated endpoints will be
removed once users are migrated onto Legacy Workflows endpoints.
}
