% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_assay-result-schemas.R
\name{get_assay_result_schemas}
\alias{get_assay_result_schemas}
\title{List assay result schemas}
\usage{
get_assay_result_schemas(schema_id = NULL, ...)
}
\arguments{
\item{schema_id}{assay_result_schemas id}

\item{...}{additional query parameters; see Details below}
}
\value{
data from \code{assay_result_schemas} endpoint
}
\description{
Schemas represent custom configuration of objects in Benchling. See this \href{https://docs.benchling.com/docs/schemas}{guide in our documentation} on how Schemas
impact our developers
}
\details{
Available query parameters:
\describe{
\item{modifiedAt}{Datetime, in RFC 3339 format. Supports the > and < operators. Time zone defaults to UTC. Restricts results to those modified in the specified range. e.g. > 2017-04-30.}
}
}
