% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_assay-runs.R
\name{get_assay_runs}
\alias{get_assay_runs}
\title{List runs}
\usage{
get_assay_runs(assay_run_id = NULL, ...)
}
\arguments{
\item{assay_run_id}{assay-runs id}

\item{...}{additional query parameters; see Details below}
}
\value{
data from \code{assay_runs} endpoint
}
\description{
Runs capture the details / parameters of a run that was performed. Results are
usually nested under a run.
}
\details{
Available query parameters:
\describe{
\item{schemaId}{ID of the assay run schema to filter by}
\item{minCreatedTime}{Filter by runs created after this unix timestamp}
\item{maxCreatedTime}{Filter by runs created before this unix timestamp}
\item{nextToken}{Token for pagination}
\item{pageSize}{Number of results to return. Defaults to 50, maximum of 100.}
\item{ids}{Comma-separated list of ids. Matches all of the provided IDs, or returns a 400 error that includes a list of which IDs are invalid.}
}
}
