% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation-output-processors.R
\name{get_automation_output_processors}
\alias{get_automation_output_processors}
\title{List non-empty Automation Output Processors}
\usage{
get_automation_output_processors(output_processor_id = NULL)
}
\arguments{
\item{output_processor_id}{automation_output_processors id}
}
\value{
data from automation_output_processors endpoint
}
\description{
Lab Automation endpoints support integration with lab instruments, and liquid
handlers to create samples or results, and capture transfers between containers
at scale.
}
