% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box-schemas.R
\name{get_box_schemas}
\alias{get_box_schemas}
\title{List box schemas}
\usage{
get_box_schemas(schema_id = NULL)
}
\arguments{
\item{schema_id}{box_schemas id}
}
\value{
data from box_schemas endpoint
}
\description{
Schemas represent custom configuration of objects in Benchling. See this \href{https://docs.benchling.com/docs/schemas}{guide in our documentation} on how Schemas impact our developers
}
