% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_boxes.R
\name{get_boxes}
\alias{get_boxes}
\title{List boxes}
\usage{
get_boxes(box_id = NULL, ...)
}
\arguments{
\item{box_id}{boxes id}

\item{...}{additional query parameters; see Details below}
}
\value{
data from \code{boxes} endpoint
}
\description{
Boxes are a structured inventory type, consisting of a grid of positions that
can each hold one container. Unlike locations, there are a maximum number of
containers that a box can hold (one per position).Boxes are all associated with
schemas, which define the type of the box (e.g. "10x10 Cryo Box") along with
the fields that are tracked and the dimensions of the box.Like all inventory,
every Box has a barcode that is unique across the registry.
}
\details{
Available query parameters:
\describe{
\item{pageSize}{Number of results to return. Defaults to 50, maximum of 100.}
\item{nextToken}{Token for pagination}
\item{schemaId}{ID of a schema. Restricts results to those of the specified schema.}
\item{schemaFields}{Schema field value. For Integer, Float, and Date type fields, supports the >= and <= operators. If present, the schemaId param must also be present. Restricts results to those with a field of whose value matches the filter.}
\item{modifiedAt}{Datetime, in RFC 3339 format. Supports the > and < operators. Time zone defaults to UTC. Restricts results to those modified in the specified range. e.g. > 2017-04-30.}
\item{name}{Name of a box. Restricts results to those with the specified name.}
\item{nameIncludes}{Name substring of a box. Restricts results to those with names that include the provided substring.}
\item{emptyPositions}{Only return boxes that have the specified number of empty positions}
\item{emptyPositions.gte}{Only return boxes that have greater-than or equal-to the specified number of empty positions.}
\item{emptyPositions.gt}{Only return boxes that have greater-than the specified number of empty positions.}
\item{emptyPositions.lte}{Only return boxes that have less-than or equal-to the specified number of empty positions.}
\item{emptyPositions.lt}{Only return boxes that have less-than the specified number of empty positions.}
\item{emptyContainers}{Only return boxes that have the specified number of empty containers (containers without contents).}
\item{emptyContainers.gte}{Only return boxes that have greater-than or equal-to the specified number of empty containers (containers without contents).}
\item{emptyContainers.gt}{Only return boxes that have greater-than the specified number of empty containers (containers without contents).}
\item{emptyContainers.lte}{Only return boxes that have less-than or equal-to the specified number of empty containers (containers without contents).}
\item{emptyContainers.lt}{Only return boxes that have less-than the specified number of empty containers (containers without contents).}
\item{ancestorStorageId}{ID of a location. Restricts results to those located in the specified inventory.}
\item{storageContentsId}{ID of a batch, entity, or entity schema. Restricts results to those that hold containers with entities or batches associated with the specified ID.}
\item{storageContentsIds}{Comma-separated list of IDs of batches or entities. Restricts results to those that hold containers with at least one of the specified batches, entities, or batches of the specified entities.}
\item{archiveReason}{Archive reason. Restricts items to those with the specified archive reason. Use "NOT_ARCHIVED" to filter for unarchived boxes. Use "ANY_ARCHIVED" to filter for archived boxes regardless of reason. Use "ANY_ARCHIVED_OR_NOT_ARCHIVED" to return items for both archived and unarchived.}
\item{ids}{Comma-separated list of ids. Matches all of the provided IDs, or returns a 400 error that includes a list of which IDs are invalid.}
\item{barcodes}{Comma-separated list of barcodes. Matches all of the provided barcodes, or returns a 400 error that includes a list of which barcodes are invalid.}
\item{names.anyOf}{Comma-separated list of names. Restricts results to those that match any of the specified names, case insensitive.  Warning - this filter can be non-performant due to case insensitivity.}
\item{names.anyOf.caseSensitive}{Comma-separated list of names. Restricts results to those that match any of the specified names, case sensitive.}
\item{creatorIds}{Comma separated list of users IDs}
}
}
