% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container-schemas.R
\name{get_container_schemas}
\alias{get_container_schemas}
\title{List container schemas}
\usage{
get_container_schemas(schema_id = NULL)
}
\arguments{
\item{schema_id}{container_schemas id}
}
\value{
data from container_schemas endpoint
}
\description{
Schemas represent custom configuration of objects in Benchling. See this \href{https://docs.benchling.com/docs/schemas}{guide in our documentation} on how Schemas impact our developers
}
