% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdowns.R
\name{get_dropdowns}
\alias{get_dropdowns}
\title{List dropdowns}
\usage{
get_dropdowns(dropdown_id = NULL)
}
\arguments{
\item{dropdown_id}{dropdowns id}
}
\value{
data from dropdowns endpoint
}
\description{
Dropdowns are registry-wide enums. Use dropdowns to standardize on spelling and
naming conventions, especially for important metadata like resistance markers.
}
