% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{extract_table}
\alias{extract_table}
\title{Extract Contents of an Entry Table}
\usage{
extract_table(tbl_contents)
}
\arguments{
\item{tbl_contents}{one row from \code{\link[=get_tables]{get_tables()}} containing the nested data}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} of the table contents
}
\description{
Extract Contents of an Entry Table
}
