% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-libraries.R
\name{get_feature_libraries}
\alias{get_feature_libraries}
\title{List Feature Libraries}
\usage{
get_feature_libraries(feature_library_id = NULL)
}
\arguments{
\item{feature_library_id}{feature_libraries id}
}
\value{
data from feature_libraries endpoint
}
\description{
Feature Libraries are collections of shared canonical patterns that can be used
to generate annotations on matching regions of DNA Sequences or AA Sequences.
}
