% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_features}
\alias{get_features}
\title{List Features}
\usage{
get_features(feature_id = NULL)
}
\arguments{
\item{feature_id}{features id}
}
\value{
data from features endpoint
}
\description{
Feature Libraries are collections of shared canonical patterns that can be used to generate annotations on matching regions of DNA Sequences or AA Sequences.
}
