% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{get_folders}
\alias{get_folders}
\title{List folders}
\usage{
get_folders(folder_id = NULL)
}
\arguments{
\item{folder_id}{folders id}
}
\value{
data from folders endpoint
}
\description{
Manage folder objects.
}
