% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_organizations.R
\name{get_organizations}
\alias{get_organizations}
\title{List organizations}
\usage{
get_organizations(organization_id = NULL, ...)
}
\arguments{
\item{organization_id}{organizations id}

\item{...}{additional query parameters; see Details below}
}
\value{
data from \code{organizations} endpoint
}
\description{
View organization objects.
}
\details{
Available query parameters:
\describe{
\item{ids}{Comma-separated list of ids. Matches all of the provided IDs, or returns a 400 error that includes a list of which IDs are invalid.}
\item{name}{Name of an organization. Restricts results to those with the specified name.}
\item{nameIncludes}{Name substring of an organization. Restricts results to those with names that include the provided substring.}
\item{names.anyOf}{Comma-separated list of names. Restricts results to those that match any of the specified names, case insensitive.  Warning - this filter can be non-performant due to case insensitivity.}
\item{names.anyOf.caseSensitive}{Comma-separated list of names. Restricts results to those that match any of the specified names, case sensitive.}
\item{modifiedAt}{Datetime, in RFC 3339 format. Supports the > and < operators. Time zone defaults to UTC. Restricts results to those modified in the specified range. e.g. > 2017-04-30.}
\item{hasMembers}{Comma-separated list of user or Benchling app IDs. Restricts results to organizations that include all the given users/apps as members.}
\item{hasAdmins}{Comma-separated list of user or Benchling app IDs. Restricts results to organizations that include all the given users/apps as admins.}
\item{nextToken}{Token for pagination}
}
}
