% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_plate-schemas.R
\name{get_plate_schemas}
\alias{get_plate_schemas}
\title{List plate schemas}
\usage{
get_plate_schemas(schema_id = NULL, ...)
}
\arguments{
\item{schema_id}{plate-schemas id}

\item{...}{additional query parameters; see Details below}
}
\value{
data from \code{plate_schemas} endpoint
}
\description{
Schemas represent custom configuration of objects in Benchling. See this \href{https://docs.benchling.com/docs/schemas}{guide in our documentation} on how Schemas
impact our developers
}
\details{
Available query parameters:
\describe{

}
}
