% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_projects}
\alias{get_projects}
\title{List projects}
\usage{
get_projects(project_id = NULL)
}
\arguments{
\item{project_id}{projects id}
}
\value{
data from projects endpoint
}
\description{
Manage project objects.
}
