% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_request-fulfillments.R
\name{get_request_fulfillments}
\alias{get_request_fulfillments}
\title{List Request Fulfillments}
\usage{
get_request_fulfillments(request_fulfillment_id = NULL, ...)
}
\arguments{
\item{request_fulfillment_id}{request_fulfillments id}

\item{...}{additional query parameters; see Details below}
}
\value{
data from \code{request_fulfillments} endpoint
}
\description{
Requests allow scientists and teams to collaborate around experimental assays
and workflows.
}
\details{
Available query parameters:
\describe{
\item{modifiedAt}{Datetime, in RFC 3339 format. Supports the > and < operators. Time zone defaults to UTC. Restricts results to those modified in the specified range. e.g. > 2017-04-30.}
}
}
