% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_requests.R
\name{get_requests}
\alias{get_requests}
\title{List requests}
\usage{
get_requests(request_id = NULL, ...)
}
\arguments{
\item{request_id}{requests id}

\item{...}{additional query parameters; see Details below}
}
\value{
data from \code{requests} endpoint
}
\description{
Requests allow scientists and teams to collaborate around experimental assays
and workflows.
}
\details{
Available query parameters:
\describe{
\item{minCreatedTime}{minimum create time (unix seconds)}
\item{maxCreatedTime}{maximum create time (unix seconds)}
}
}
