% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_workflow-stages.R
\name{get_workflow_stages}
\alias{get_workflow_stages}
\title{Legacy Workflows (deprecated)}
\usage{
get_workflow_stages(stage_id = NULL, ...)
}
\arguments{
\item{stage_id}{workflow_stages id}

\item{...}{(ignored)}
}
\value{
data from \code{workflow_stages} endpoint
}
\description{
Please use endpoints for Legacy Workflows. These deprecated endpoints will be
removed once users are migrated onto Legacy Workflows endpoints.
}
