% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_workflow-task-groups.R
\name{get_workflow_task_groups}
\alias{get_workflow_task_groups}
\title{List workflow task groups}
\usage{
get_workflow_task_groups(workflow_task_group_id = NULL, ...)
}
\arguments{
\item{workflow_task_group_id}{workflow_task_groups id}

\item{...}{additional query parameters; see Details below}
}
\value{
data from \code{workflow_task_groups} endpoint
}
\description{
Workflow task groups are groups of workflow tasks of the same schema
}
\details{
Available query parameters:
\describe{
\item{ids}{Comma separated list of workflow task group IDs}
\item{schemaId}{The workflow task schema ID of tasks in this task group}
\item{folderId}{A folder ID}
\item{projectId}{A project ID}
\item{mentionedIn}{A comma separated list entry IDs}
\item{watcherIds}{Comma separated list of user IDs or "null"}
\item{executionTypes}{Comma separated list of workflow execution types. Acceptable execution types are "DIRECT" and "ENTRY"}
\item{responsibleTeamIds}{Comma separated list of team IDs or "null"}
\item{statusIds.anyOf}{Commas separated list of Status ids. Returns workflows where at least one task is of one of the provided statuses.}
\item{statusIds.noneOf}{Commas separated list of Status ids. Returns workflows where none of the tasks are of any of the provided statuses.}
\item{statusIds.only}{Commas separated list of Status ids. Returns workflows where all of the tasks are of one of the provided statuses.}
\item{name}{The name of the workflow task group}
\item{nameIncludes}{Part of the name of the workflow task group}
\item{creatorIds}{Comma separated list of user IDs.}
\item{modifiedAt}{Datetime, in RFC 3339 format. Supports the > and < operators. Time zone defaults to UTC. Restricts results to those modified in the specified range. e.g. > 2017-04-30.}
\item{displayIds}{Comma-separated list of Workflow Display IDs.}
\item{archiveReason}{Archive reason. Restricts items to those with the specified archive reason. Use "NOT_ARCHIVED" to filter for unarchived workflow task groups. Use "ANY_ARCHIVED" to filter for archived workflow task groups regardless of reason. Use "ANY_ARCHIVED_OR_NOT_ARCHIVED" to return items for both archived and unarchived.}
}
}
